etl/requirements.txt
SQLAlchemy
psycopg2-binary
pymongo
vaderSentiment
pandas
